+++ packages/browser/src/login/oidc/__mocks__/IssuerConfigFetcher.ts
@@ -32,6 +32,7 @@ export const IssuerConfigFetcherFetchConfigResponse: IIssuerConfig = {
   subjectTypesSupported: [],
   claimsSupported: [],
   grantTypesSupported: ["refresh_token"],
+  idTokenSigningAlgValuesSupported: ["RS256"],
 };
 
 export const IssuerConfigFetcherMock: jest.Mocked<IIssuerConfigFetcher> = {